backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.31.1
pydantic==2.11.0
python-dotenv==1.0.1
python-multipart==0.0.21
PyYAML==6.0.1
agent-framework-core==1.0.0b260130
agent-framework-github-copilot==1.0.0b260130
agent-framework-claude==1.0.0b260130
agent-framework-azure-ai==1.0.0b260130
azure-ai-inference==1.0.0b5
azure-core
